+++ src/index.ts
@@ -1,6 +1,6 @@
 import { RPReporter } from './reporter';
 export { ReportingApi } from './reportingApi';
 export { LOG_LEVELS, STATUSES } from './constants';
-export { RPTestInfo } from './constants';
+export { ReportPortalConfig } from './models';
 
 export default RPReporter;