+++ src/reporter.ts
@@ -41,7 +41,6 @@ import {
   getAgentInfo,
   getAttachments,
   getCodeRef,
-  getDescription,
   getSystemAttributes,
   isErrorLog,
   isFalse,
@@ -346,8 +345,6 @@ export class RPReporter implements Reporter {
         ...(status && { status }),
         ...(logs && { logs }), // TODO: may be send it on suite start
       });
-
-      this.suitesInfo.delete(currentSuiteTitle);
     }
 
     return projectName;
@@ -455,22 +452,20 @@ export class RPReporter implements Reporter {
       return Promise.resolve();
     }
 
-    const additionalInfo = getAdditionalInfo(test);
-
     const {
-      id: testItemId,
       attributes,
       description,
-      testCaseId,
       status: predefinedStatus,
-    } = savedTestItem;
+      testCaseId,
+    } = getAdditionalInfo(test);
+
+    const { id: testItemId } = savedTestItem;
     let withoutIssue;
-    let testDescription = getDescription(description, additionalInfo.description);
+    let testDescription = description;
 
     const calculatedStatus = calculateRpStatus(test.outcome(), result.status, test.annotations);
-    const status = predefinedStatus || additionalInfo.status || calculatedStatus;
-    const calculatedTestCaseId = testCaseId || additionalInfo.testCaseId;
-    const mergedAttributes = additionalInfo.attributes.concat(attributes ?? []);
+    const status = predefinedStatus || calculatedStatus;
+
     if (status === STATUSES.SKIPPED) {
       withoutIssue = isFalse(this.config.skippedIssue);
     }
@@ -497,9 +492,7 @@ export class RPReporter implements Reporter {
         level: LOG_LEVELS.ERROR,
         message: stacktrace,
       });
-      testDescription = getDescription(description, additionalInfo.description).concat(
-        `\n\`\`\`error\n${stacktrace}\n\`\`\``,
-      );
+      testDescription = description.concat(`\n\`\`\`error\n${stacktrace}\n\`\`\``);
     }
 
     [...this.nestedSteps.entries()].forEach(([key, value]) => {
@@ -521,9 +514,9 @@ export class RPReporter implements Reporter {
       endTime: this.client.helpers.now(),
       status,
       ...(withoutIssue && { issue: { issueType: 'NOT_ISSUE' } }),
-      ...(mergedAttributes.length !== 0 && { attributes: mergedAttributes }),
+      ...(attributes.length !== 0 && { attributes }),
       ...(testDescription && { description: testDescription }),
-      ...(calculatedTestCaseId && { testCaseId: calculatedTestCaseId }),
+      ...(testCaseId && { testCaseId }),
     };
     const { promise } = this.client.finishTestItem(testItemId, finishTestItemObj);